Refactor redis storage and sender
